refactor: unblock and extract the certificate download endpoints (#667)
The decomposition stalled here. Both download classes call
_validate_domain_path, which lived in resources.py alongside thirteen other
callers, so a module of their own could not use it without importing the file
that imports them back.

path_validation.py is that helper's shared home. Nothing about the rules
changes — the same three checks in the same order, and resources.py re-exports
_validate_domain_path and _DOMAIN_RE under their old names so the thirteen
remaining call sites and the existing tests are untouched.

With that out of the way, DownloadCertificate and DownloadCertificateFile move
into resources_downloads.py. Closure complexity 249 -> 196; resources.py
2023 -> 1658 lines. Since this work began: 559 -> 196, and 4040 -> 1658.

_PRIVATE_KEY_FILES and _PUBLIC_DOWNLOAD_FILES travel with them. They are the
access rule — which downloads require operator rather than viewer — and they
belong beside the endpoints that enforce it rather than remaining loose closure
state.

That rule is now tested rather than assumed, and the two endpoints are tested
SEPARATELY. A first version covered only DownloadCertificateFile; mutating the
gate in DownloadCertificate changed nothing and every test stayed green, which
is precisely the failure this kind of move can cause. Each gate is now
mutation-verified on its own, each with a control proving a viewer can still
reach the public files — an endpoint that refused everything would satisfy the
denial assertions while being broken.

_privkey_to_pkcs1 is passed in as an argument rather than moved: its tests
import it from resources.py, and taking it as a parameter avoids an import
cycle without dragging a second helper into the same change.
